+++ examples/main/variables.tf
@@ -1,0 +1,9 @@
+variable "myresource_id_1" {
+  description = "Resource ID 1"
+  type        = string
+}
+
+variable "myresource_id_2" {
+  description = "Resource ID 2"
+  type        = string
+}